refactor: rename systick.[ch] to systimer.[ch]
