+++ raspberrypi/modutime.c
@@ -32,7 +32,7 @@
 #include "py/obj.h"
 #include "lib/timeutils/timeutils.h"
 #include "extmod/utime_mphal.h"
-#include "systick.h"
+#include "systimer.h"
 
 /// \module time - time related functions
 ///
