+++ raspberrypi/mphalport.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <unistd.h>
 #include "py/mpconfig.h"
-#include "systick.h"
+#include "systimer.h"
 #include "uart-qemu.h"
 #include "usb.h"
